+++ lib/ring.ml
@@ -329,10 +329,10 @@ module type S = sig
   module Reader: READABLE
   module Writer: WRITABLE
 
-  val write: Cstruct.t -> string -> int -> int -> int
+  val write: Cstruct.t -> bytes -> int -> int -> int
   val read: Cstruct.t -> bytes -> int -> int -> int
 
-  val unsafe_write: Cstruct.t -> string -> int -> int -> int
+  val unsafe_write: Cstruct.t -> bytes -> int -> int -> int
   val unsafe_read: Cstruct.t -> bytes -> int -> int -> int
 end
 
@@ -414,7 +414,7 @@ module Pipe(RW: RW) = struct
     let seq, frag = Writer.write t in
     let free_space = Cstruct.len frag in
     let can_write = min len free_space in
-    Cstruct.blit_from_string buf ofs frag 0 can_write;
+    Cstruct.blit_from_bytes buf ofs frag 0 can_write;
     Writer.advance t Int32.(add seq (of_int can_write));
     can_write