+++ CHANGES.md
@@ -1,6 +1,6 @@
 ## 3.0.0 (2017-11-05)
 
-* Update to use `bytes` in the `read` call, to work with OCaml 4.06 and
+* Update to use `bytes` in the `read` and `write`, to work with OCaml 4.06 and
   `-safe-string`
 
 ## 2.0.1 (2017-06-07)